Add NativeAudioRecordReader to read audio data (AAC, MP3, WAV, etc) with FFmpeg
