+++ datavec-data/datavec-data-audio/src/test/java/org/datavec/audio/AudioReaderTest.java
@@ -1,0 +1,63 @@
+/*-
+ *  * Copyright 2017 Skymind, Inc.
+ *  *
+ *  *    Licensed under the Apache License, Version 2.0 (the "License");
+ *  *    you may not use this file except in compliance with the License.
+ *  *    You may obtain a copy of the License at
+ *  *
+ *  *        http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *    Unless required by applicable law or agreed to in writing, software
+ *  *    distributed under the License is distributed on an "AS IS" BASIS,
+ *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *    See the License for the specific language governing permissions and
+ *  *    limitations under the License.
+ */
+
+package org.datavec.audio;
+
+import static org.junit.Assert.*;
+
+import org.bytedeco.javacv.FFmpegFrameRecorder;
+import org.bytedeco.javacv.Frame;
+import org.datavec.api.records.reader.RecordReader;
+import org.datavec.api.split.FileSplit;
+import org.datavec.api.writable.Writable;
+import org.datavec.audio.recordreader.NativeAudioRecordReader;
+import org.junit.Ignore;
+import org.junit.Test;
+
+import java.io.File;
+import java.nio.ShortBuffer;
+import java.util.List;
+
+import static org.bytedeco.javacpp.avcodec.*;
+
+/**
+ * @author saudet
+ */
+public class AudioReaderTest {
+    @Ignore
+    @Test
+    public void testNativeAudioReader() throws Exception {
+        File tempFile = File.createTempFile("testNativeAudioReader", ".ogg");
+        FFmpegFrameRecorder recorder = new FFmpegFrameRecorder(tempFile, 2);
+        recorder.setAudioCodec(AV_CODEC_ID_VORBIS);
+        recorder.setSampleRate(44100);
+        recorder.start();
+        Frame audioFrame = new Frame();
+        ShortBuffer audioBuffer = ShortBuffer.allocate(64 * 1024);
+        audioFrame.sampleRate = 44100;
+        audioFrame.audioChannels = 2;
+        audioFrame.samples = new ShortBuffer[] {audioBuffer};
+        recorder.record(audioFrame);
+        recorder.stop();
+        recorder.release();
+
+        RecordReader reader = new NativeAudioRecordReader();
+        reader.initialize(new FileSplit(tempFile));
+        assertTrue(reader.hasNext());
+        List<Writable> record = reader.next();
+        assertEquals(audioBuffer.limit(), record.size());
+    }
+}
